General-purpose multi-level paging is basically functional except for one issue: Nothing renders (in the page2 test) until you increase the view distance out to the lowest LOD. After that, you can decrease the view distance and watch higher LODs page in.
There are some other issues too, for example sometimes it gets "stuck" at the lowest LOD, or sometimes it sets the node masks incorrectly and some boxes don't render. I'll be working on all these issues over the summer.

Once this is working well, I'll move the preprocess op out of the page2 example and into the base lfx library.